DataAnalystProject/requirements.txt
# Data Science Hub Requirements

# Core
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.3.0

# Advanced ML (AutoML)
optuna>=3.4.0
xgboost>=2.0.0
lightgbm>=4.1.0
catboost>=1.2.0

# Excel/Documents
openpyxl>=3.1.0
python-docx>=0.8.11

# PDF
reportlab>=4.0.0

# Model Interpretability
shap>=0.43.0

# Data formats
pyarrow>=14.0.0
fastparquet>=2023.10.0

# Statistics
scipy>=1.11.0

# API (optional)
# fastapi>=0.104.0
# uvicorn>=0.24.0

# Advanced Preprocessing
imbalanced-learn>=0.11.0
sqlalchemy>=2.0.0

